# backend/docs/openapi.yaml
openapi: 3.0.3
info:
  title: AI Agency API
  description: |
    API da plataforma AI Agency para gerenciamento e geração de conteúdo usando múltiplos provedores de IA.
    
    ## Recursos Principais
    - Gerenciamento de prompts e templates
    - Integração com múltiplos provedores (OpenAI, Anthropic, Cohere)
    - Sistema de filas e rate limiting
    - Monitoramento e métricas
  version: 1.0.0
  contact:
    name: Time Backend
    email: backend@aiagency.com
servers:
  - url: http://localhost:8000
    description: Ambiente de Desenvolvimento
  - url: https://staging.aiagency.com
    description: Ambiente de Staging
  - url: https://api.aiagency.com
    description: Ambiente de Produção
...
paths:
...
  /templates:
    get:
      tags:
        - templates
      summary: Listar templates
      description: Retorna a lista de templates disponíveis
      operationId: listTemplates
      security:
        - bearerAuth: []
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: per_page
          in: query
          schema:
            type: integer
            default: 10
      responses:
        '200':
          description: Lista de templates
          content:
            application/json:
              schema:
                type: object
                properties:
                  items:
                    type: array
                    items:
                      $ref: '#/components/schemas/Template'
                  total:
                    type: integer
                  page:
                    type: integer
                  per_page:
                    type: integer
    
    post:
      tags:
        - templates
      summary: Criar template
      description: Cria um novo template
      operationId: createTemplate
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TemplateCreate'
      responses:
        '201':
          description: Template criado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Template'
...
components:
  schemas:
    Template:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        description:
          type: string
        prompt:
          type: string
        variables:
          type: array
          items:
            type: object
            properties:
              name:
                type: string
              type:
                type: string
                enum: [text, number, boolean, options]
              required:
                type: boolean
              default:
                type: string
              options:
                type: array
                items:
                  type: string
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    TemplateCreate:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
        prompt:
          type: string
        variables:
          type: array
          items:
            type: object
            properties:
              name:
                type: string
              type:
                type: string
                enum: [text, number, boolean, options]
              required:
                type: boolean
              default:
                type: string
              options:
                type: array
                items:
                  type: string
      required:
        - name
        - prompt
        - variables
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT 